+++ src/libAtomVM/CMakeLists.txt
@@ -50,6 +50,7 @@ set(HEADER_FILES
     port.h
     refc_binary.h
     scheduler.h
+    stacktrace.h
     sys.h
     term_typedef.h
     term.h
@@ -81,6 +82,7 @@ set(SOURCE_FILES
     port.c
     refc_binary.c
     scheduler.c
+    stacktrace.c
     term.c
     timer_wheel.c
     valueshashtable.c
